+++ src/main/java/DAO/AuthorDao.java
@@ -7,7 +7,7 @@ public interface AuthorDao extends BaseDao <Author> {
     int addEntity (Author author);
     int addTranslation (Author author);
     List<Author> getAll (String lang);
-    int deleteById (int id);
-    int deleteByIdLang (int id, String lang);
+    int deleteById (long id);
+    int deleteByIdLang (long id, String lang);
     List <Integer> searchAuthors (String search);
 }
