+++ src/main/java/DAO/FormatDao.java
@@ -7,5 +7,5 @@ public interface FormatDao extends BaseDao <Format>{
 
     int addEntity (Format format);
     List<Format> getAll (String lang);
-    int deleteById (int id);
+    int deleteById (long id);
 }
