+++ src/main/java/DAO/OrderDao.java
@@ -7,7 +7,7 @@ public interface OrderDao extends BaseDao <Order> {
 
     int addEntity(Order order);
     List <Order> getAll ();
-    int updateStatus (int orderId, int statusId);
-    Order getOrderById (int id);
-    List <Order> getOrdersByUserId(int userId);
+    int updateStatus (long orderId, long statusId);
+    Order getOrderById (long id);
+    List <Order> getOrdersByUserId(long userId);
 }
