+++ src/main/java/DAO/UserDao.java
@@ -6,13 +6,13 @@ import java.util.List;
 public interface UserDao extends BaseDao <User> {
 
     int addEntity (User user);
-    int addAddress (int id, String address);
+    int addAddress (long id, String address);
     List <User> getAll();
     boolean isAdmin (String email);
     int validateUser (String email, String password);
-    User getUser (int id);
+    User getUser (long id);
     boolean userExists (String email);
-    int blockUser(int id, boolean status);
-    int changePassword (int id, String oldPass, String newPass);
+    int blockUser(long id, boolean status);
+    int changePassword (long id, String oldPass, String newPass);
 
 }
